Code cleanup and stupidity fixing.
