+++ src/main/java/pro/i_it/learn/server/servlet/InitializeListener.java
@@ -4,6 +4,7 @@ import com.corundumstudio.socketio.Configuration;
 import com.corundumstudio.socketio.SocketIOServer;
 import pro.i_it.learn.server.model.MessageModel;
 import pro.i_it.learn.server.module.ModuleManager;
+import pro.i_it.learn.server.module.socketIOManager.pansmile.SocketIOManager;
 
 import javax.servlet.ServletContextEvent;
 import javax.servlet.ServletContextListener;
@@ -14,44 +15,33 @@ import javax.servlet.annotation.WebListener;
  */
 @WebListener
 public class InitializeListener implements ServletContextListener {
-    private SocketIOServer server;
+    private SocketIOManager socketIOManager = (SocketIOManager) ModuleManager.getModuleManager().getSocketIOManager();
 
 
     @Override
     public void contextInitialized(ServletContextEvent servletContextEvent) {
         System.out.println("startup");
         try {
-            initServer();
-            server.start();
+            socketIOManager.startSocketIOServer();
             System.out.println("SERVER STARTED!");
         } catch (Exception e0) {
             e0.fillInStackTrace();
         }
 
+
+
     }
 
     @Override
     public void contextDestroyed(ServletContextEvent servletContextEvent) {
-        server.stop();
-        System.out.println("SERVER STOPPED!");
+        try {
+            socketIOManager.stopSocketIOServer();
+            System.out.println("SERVER STOPPED!");
+        } catch (Exception e0) {
+            e0.fillInStackTrace();
+        }
         System.out.println("shutdown");
 
     }
 
-   private void initServer() throws Exception {
-        if (server == null) {
-            Configuration config = new Configuration();
-            config.setPort(9090);
-            config.setHostname("localhost");
-            server = new SocketIOServer(config);
-            System.out.println("SERVER CREATED");
-            server.addEventListener("MESSAGE", String.class,
-                    (client, message, ackRequest) -> {
-                        System.out.println("Client said: " + message);
-                        ModuleManager.getModuleManager().getChatManager().inputMessage(
-                                new MessageModel(client.getNamespace().getName(), message));
-                    });
-
-        }
-    }
 }
